# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bff_calendar",
    "version": "1.0.0"
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CalendarAssignee": {
        "type": "object",
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ],
            "description": "Identifiant unique de la personne",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Nom complet affichable",
            "example": "Security Admin"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Adresse email",
            "example": "security-admin@mairie360.fr"
          },
          "role": {
            "type": "string",
            "description": "Fonction ou rôle métier",
            "example": "Admin"
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri",
            "description": "URL de l'image de profil",
            "example": "https://mairie360.fr/avatar.png"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "BAD_REQUEST",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "PAYLOAD_TOO_LARGE",
                  "UNPROCESSABLE_ENTITY",
                  "TOO_MANY_REQUESTS",
                  "INTERNAL_ERROR",
                  "BAD_GATEWAY",
                  "SERVICE_UNAVAILABLE",
                  "GATEWAY_TIMEOUT"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            },
            "required": [
              "code",
              "message",
              "details"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/calendar/assignees": {
      "get": {
        "tags": [
          "Calendar"
        ],
        "summary": "Récupère le référentiel des personnes assignables",
        "description": "Charge la liste complète des personnes pouvant être assignées à un événement",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "example": "2030-06-01"
            },
            "description": "Date de début au format YYYY-MM-DD (défaut : premier jour du mois courant)"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "example": "2030-06-30"
            },
            "description": "Date de fin au format YYYY-MM-DD (défaut : dernier jour du mois courant)"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des personnes assignables",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CalendarAssignee"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters: dates not in the YYYY-MM-DD format, from after to, or more than 1096 days apart",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Calendar API or Core API is unavailable or failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Calendar API, Core API or JWT_SECRET is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}